+++ src/learn/spring/model/PersonMapper.java
@@ -10,7 +10,7 @@ public class PersonMapper implements RowMapper<Person> {
 	public Person mapRow(ResultSet resultSet, int i) throws SQLException {
 
 		Person person = new Person();
-		person.setId(resultSet.getLong("id"));
+		person.setId(resultSet.getInt("id"));
 		person.setFirstName(resultSet.getString("first_name"));
 		person.setLastName(resultSet.getString("last_name"));
 		person.setAge(resultSet.getInt("age"));
